Move mouse query to helper
