+++ meerk40t/gui/wxmmain.py
@@ -226,54 +226,12 @@ class MeerK40t(MWindow):
         self.tips_at_startup()
         self.parametric_info = None
 
-        # Look at window elements we are hovering over
-        # to establish the online help functionality
-        self.context.kernel.add_job(run=self.mouse_query, name="helper-check", interval=0.5, run_main=True)
-        # Switched to kernel to avoid 0xC0000005 crash in windows.
-        # self.timer = wx.Timer(self, id=wx.ID_ANY)
-        # self.Bind(wx.EVT_TIMER, self.mouse_query, self.timer)
-        # self.timer.Start(500, wx.TIMER_CONTINUOUS)
-        self._last_help_info = ""
-        self._last_help_section = ""
-
-    def mouse_query(self, event=None):
-        """
-            This routine looks periodically (every 0.5 seconds)
-            at the window ie control under the mouse cursor.
-            It will examine the window if it contains a tooltip text and
-            will display this in a textbox in this panel.
-            Additionally, it will read the associated HelpText of the control
-            (or its parent if the control does not have any) to construct a
-            Wiki page on GitHub to open an associated online help page.
-        """
-        wind, pos = wx.FindWindowAtPointer()
-        if wind is not None:
-            info = ""
-            if hasattr(wind, "GetToolTipText"):
-                info = wind.GetToolTipText()
-            section = ""
-            if hasattr(wind, "GetHelpText"):
-                win = wind
-                while win is not None:
-                    section = win.GetHelpText()
-                    if section:
-                        break
-                    win = win.GetParent()
-                if section is None or section == "":
-                    section = "GUI"
-            if info != self._last_help_info or section != self._last_help_section:
-                self._last_help_info = info
-                self._last_help_section = section
-                # Inform HelperWindow about a new content
-                self.context.signal("helpinfo", info, section)
-
     def tips_at_startup(self):
         self.context.setting(bool, "show_tips", True)
         if self.context.show_tips:
             self.context("window open Tips\n")
         else:
             self.context("window close Tips\n")
-
 
     def update_check_at_startup(self):
         if self.context.update_check == 0:
@@ -3518,7 +3476,8 @@ class MeerK40t(MWindow):
             )
 
         def online_help(event):
-            sect = self._last_help_section
+            origin, message = self.context.last_signal("helpinfo")
+            _, sect = message
             if sect is None or sect == "":
                 sect = "GUI"
             sect = sect.upper()
